made master ready for NetworkModule, fixed wrapper mongodb downloader

+++ master/src/main/java/eu/mcone/cloud/master/server/StaticServerManager.java
@@ -36,7 +36,7 @@ public class StaticServerManager {
                                     "",
                                     entry.getInteger("max"),
                                     0,
-                                    entry.getInteger("ram"),
+                                    entry.getLong("ram"),
                                     true,
                                     ServerVersion.valueOf(entry.getString("version")),
                                     "{\"plugins\":[], \"worlds\":[], \"gamemodeType\":[], \"configs\":[]}"
